pkg/pgreplay/integration
Add integration tests for the entire replay flow, covering the parsing
of logs through to the result we expect in the database. This commit
also fixes a couple of bugs that were surfaced by the addition of a
proper test suite.

One subtle but important bug was caused by a side effect of us
incorrectly signalling that we'd parsed a DETAIL parameter log-line. As
we returned nil for these entries, it was possible for subsequent
DETAILs to override the query parameters of the previously unbound
query, clobbering whatever had been bound. We now have a special
BoundExecuteItem which solves this problem, but we should re-evaluate
the current approch of modelling query types.