Refactor to simplify
Remove the need to perform type switches by extending a parsed item's
interface to know how to handle itself. This means we can better
leverage the type system and remove the possibility that we can put
anything into the channel that can't take an appropriate action on the
connection.

Have the Database manage creating connections, and don't pass the
WaitGroup away from the main Database loop. The connection can terminate
itself in response to a Disconnect now, simplifying some of the
termination logic.