Execute can't Handle
It's quite subtle that an Execute log isn't sufficient to resolve how to
run the associated query. This commit removes the Handle method from
Execute to clearly enforce that Execute should not be considered a
complete parsed Item, forcing any code that wants to supply an Execute
with the parameters it should be bound with.