Reimplement Pending
This is a bit of a crap-shoot. I don't want to introduce locks around
accessing the map because this may slow things down, but I definitely
want to have the pending method and that requires reading the map while
risking a concurrent write. I think we're highly unlikely to see any
problems but want to call it out that Pending should be suspect if we
begin to see segfaults.