Change implementation of streamer
The previous streamer assumed pushing down a channel was immediate,
which may not be correct. This streamer uses a simulated clock to push
items as soon as they should have been sent. If the channel blocks for a
moment, the streamer will push a succession of items until the timestamp
catches up.